backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2